byte slice copying avoids bounds checks when it is convenient
Using the copy() builtin function provides better performance than a
naive for loop when copying long slices by eliminating redundant bounds
checks.  However, a "copy" tag in snappy-encoded content can have a
length greater than its backword offset (section 2.2 of the formal
specification).  The implementation of the copy() builtin doesn't behave
desirably when copying such a range.  So, in such cases, the old naive
for loop is used for a fallback in favor of more complex logic as it
become hard to efficiently determine the effectiveness of copying
smaller segments using the builtin function.

Specification at the time of authoring:

    https://github.com/google/snappy/blob/b61134bc0a6a904b41522b4e5c9e80874c730cef/format_description.txt